+++ peyotl/phylografter/sync.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 from peyotl.phylografter.nexson_workaround import workaround_phylografter_nexson
-from peyotl.utility.io import open_for_group_write
+from peyotl.utility.input_output import open_for_group_write
 from peyotl.nexson_syntax import read_as_json
 from peyotl.utility import get_logger
 import datetime
